release-files: bugfix, code cleanup.
The ReleaseFilesEvent._find_rpms() method first looks for
'<product-name>-release' RPM, and if it doesn't find it, it looks for
'*release*' and '*release-notes* RPMs'. There is a potential bug here in
the case where the product name contains the substring "release". So
now, I have added an nglob for '*<product-name>*'.