logos-rpm: Complete rewrite of logos RPM generation.
The logos-rpm event now makes user customization of the output images
really easy. There is a distro-specific config file located at
share/logos-rpm/{redhat5,centos5,fedora9}.xml. If no distro-specific
config file is found, a file called "default.xml" is searched for in
the share paths and used.

The config file can use macros and XIncludes similar to the distro
definition file.

This config file defines its precedence. It defaults to 1 and is used
to order multiple config files. If two files have the same precedence
the second config file's data overrides the first one's.

It also contains <file/> and<image/> elements. The <file/> element
simply copies a file and installs it to the location defined by the
element. The config file writer can choose to have files specific to
XWindow Types and Anaconda Versions; <file/>s can also be removed for
logos RPMs generated for distributions after a specific anaconda
version.

The <image/> elements are of three types: <create/>, <source/>, or
<remove/>. The <create/> element can be used to define the creation of
an image. The <source/> image is the same as <file/> elements and they
can be used interchangeably. The <remove/> element is also the same as
the one for <file/> elements in that it removes images after a certain
anaconda version.