Using IOMixin.sync_input() to sync files in the publish event.
The publish event was linking files from the build directory to the
publish directory itself, instead of using `self.io.sync_input()'.
This was being done so that the publish directory can be cleaned up as
and if needed. Now that the `PublishEvent.apply()' method has been
updated to take care of the cleaning of the publish directory, we can
use the `self.io.sync_input(link=True)' again.