Chdir'ing into the config file's folder at a later point in time.
Previously, we were chdir'ing into the distrodef file's folder soon after
reading the distrodef file. This breaks the setting up of the share
paths because if a share path specified at the command line is
relative, it is made relative to the distrodef file's directory. Now,
we chdir into the distrodef's directory after setting up all the Event
objects' default values.