The 'pkglist' event prints out warnings for user-required packages.
Previously, warnings were printed out for obsolete packages for all
packages that ended up in the comps file.  This became too verbose at
times because there are packages in the "core" group that have been
renamed as well.  The user has no control over the packages in the
"core" group and cannot modify it.  Now, only packages that are
explicitly added to the comps file by the user have warnings printed
out for them.