Being smarter about package updates; code cleanup.
The IDepsolver.iupdate() method didn't consider flags for requirements
of packages.  It just checked the flags of requirements when the
requirement's name matched the name of the package with the update.
This wasn't sufficient all the time.  For instance, the
python-urlgrabber package provides urlgrabber, and other packages
require urlgrabber (usually) instead of python-urlgrabber.