The <package/> element is not supported for 'logos' and 'release-files'.
The <package/> element was a way to specify the RPM to use as the
logos RPM (for the logos event) and the RPM(s) to use as the release
RPM(s) (for the release-files event).  A better way to accomplish this
is to have pkglist set up a control variable with the names of the
logos RPM and release RPM(s).  This changeset is a step in that
direction.  Look at Bug 309 for more details.