The location of files that end up in the custom RPMs has changed.
Before this changeset, the build/ folder used to be the folder in
which files were to be laid out based on the locations they should be
installed at.  This becomes icky when we start adding scripts and
triggers to the build/ folder.  Now, all the files that are to be
installed by the RPM need to be placed at build/source.  This gets rid
of some of the nglobs we have while computing the list of data_files.