Temporary fix to missing spin.conf files.
If a spin.conf file is not found, validation fails when trying to
print out a log message.  In the case of a missing spin.conf file, we
create a StringIO object with the contents "<spin/>".  This object has
not file attribute which was being accessed while trying to print out
a message.  For now, just printing out 'spin.conf' as the main
config's file.  This is not always true because you can specify your
own main config file and call it whatever.  I am working on a fix.