The <remove> element's usage implies that image/file should be removed.
If you have a <remove> element in any of the <image> or <file>
elements, we assume that it should be removed.  Having a <remove>
element's text value set to 'False' (or some other accepted variable)
is meaningless.  Therefore, <remove/> elements cannot have any
children and if present mean that they the image or file it is
associated with should be removed.