The SpinRepo._pkg_filter() method now supports globs.
If the repository had a list of included packages associated with it
and some of them were globs, when a package (which matched one of the
globs) was tested, the method would incorrectly return False.

To get around this problem, we now convert the list of globs into a
regular expression, and then test.