Added assert_file_has_content() function to errors module.
The assert_file_has_content() function makes sure that the file is
readable and that it has content.

The assert_file_readable() method just makes sure that the file is
readable and is used by assert_file_has_content() makes use of that to
check that it is.

Made changes in all the Event classes to use the correct function.