Fixed Bug 362.  Added [checksums] section to treeinfo file.
Events that want some of their output files' checksum written out to
the treeinfo file need to modify the 'treeinfo-checksums' control
variable by adding the path to their software store and the relative
path to the file from the software store (as a tuple) to it.

The treeinfo computes a list of files based on the tuples and marks
them as input files.  It then -- in its run() method -- iterates over
this list of tuples, and computes the checksum of the file and then
writes it out to the treeinfo file.