Added support for conditional packages in depsolve event.
The depsolve event was ignoring the conditional packages in the
comps.xml.  Now, the comps event sets up a
'comps-conditional-packages' cvar that is a list of (conditional
package, requires) tuples.

The depsolve event reads this cvar and if a requirement of a
conditional package is in the list of mandatory, default, and optional
packages and their deps, the conditional package and its deps get
added to the list.

Also, renamed the 'pkglist-mandatory-packages' cvar to
'pkglist-install-packages'.  The vm events now look at the
pkglist-install-packages variable instead of the 'pkglist' variabl,
and run if the variable changes.