The per-group defaults and optional attributes are made use of while computing 'pkglist-install-packages'.
In the <packages/> element's <group/> child elements, the @default and
@optional attributes can be used to specify whether the default and
optional packages of that group should be used or not.  The depsolve
event makes use of these attributes while computing the
'pkglist-install-packages' cvar.

Also, the kickstart event, while writing out the kickstart file writes
out the `--nodefaults' and `--optional' arguments for each group as
needed.