config: backed out input validation; modified postun event
This fix makes three changes to the config event.

The first is to comment out the recently added code to implement
friendly error messages for input validation. This
code made use of assert_file_readable, which operates, as the name
implies, at a file level. The unintended consequence is that it was
blocking the use of folders within config/files elements. Need to come
up with a better fix, but for now we will simply back it out.

The second is to modify the postun script to remove obsolete files when
the config rpm is updated. The previous code was only removing files
when the rpm was removed entirely. In addition, the postun script
now removes obsolete dirs from the /usr/share/$name/files folder. More
work is needed to have postun remove obsolete dirs created by the post
script.

The third is to bump the version.