better .rpmsave handling for config rpm
Previous, the config rpm created a backup of an existing file (with the
.rpmsave file extension) each time a file changed, whether by user mods
to the file on disk or by admin mods to the file in the distribution.
With this commit, the config rpm will only create rpmsave files in the
former case.

systembuilder/modules/core/rpmbuild/config.py
* added a _generate_files_checksums method that includes a file
containing checksums in the rpm
* added a pre scriptlet that copies the checksums file (from the prior
rpm install) to the same file with a .prev extension
* added logic to the post script that compares checksums and creates an
.rpmsave file only if the file on disk does not match the same file in
either the current or previous rpm.
* added logic to the postun script to remove the checksum file with the
.prev extension