config: adding preserve=all to file copy in post scriptlet
Previously the post scriptlet of the config rpm was copying files to the
destination folder without bothering to preserve file attributes (mode,
timestamp, etc.)  This was causing problems for some applications that
test for file attribute changes (esp. timestamp). For example, if the
/etc/aliases file on disk has a timestamp newer than the timestamp of
the previously compiled (using newalises) file, sendmail gives a
warning. The fix is to use the 'preserve=all' parameter when copying.

systembuilder/modules/core/rpmbuild.config
* bumped version number
* added 'preserve=all' attribute to the post scriptlet copy command