removing optional sources event
While the sources event works currently, it relies on manually reading the primary repo file as, in the past at least, not all os distros have run createrepo on their SRPM folders. We want to get away from manually reading the primary repo file for performance reasons, for one, and because of a segfault with our current implementation in python 2.6, for another.

Since sources is an optional event, the path of least resistance is to remove it, and additionally proceed with eliminating all cases where we manually read the primary repo file.