performance enhancement: eliminated repo.pkgsfile
Previously, the systemstudio repos event read the primary.xml file for each repository to get a list of packages.  This list was used later in the repos event to set the anaconda-version variable. It was also used by the download event to determine the repository for downloading packages. Reading the primary.xml, saving content to a separate repo.pkgsfile, and rereading the repo.pkgsfile, were timeconsuming tasks, especially for large repositories. This was all the more a shame because the primary.xml file was read again, a second time, by the depsolver event, more accurately by yum, as the repositories were being setup for dependency resolution.

This commit eliminates the read of primary.xml files by the repos event. Now the anaconda-version is set by reading the OS distribution name and version from the treeinfo file of the installer repository, and mapping this to an anaconda-version. In addition, the depsolver event now provides the repository where packages are located in the pkglist cvar datastructure, and the download event uses this variable, rather than repo.pkgsfiles to determine package download location.