Removing keys for empty values in rpms-by-repoid
The gpgkey event uses cvars['rpms-by-repoid'] for displaying status. It assumes that every repo represented in the dictionary contains a list of rpms. This changed recently when the download event began to include empty values. The fix is for the download event to filter out empty items before populating the dictionary. This commit also modifies the code a bit to make it more efficient.