getting createrepo version from createrepo rather than from rpm
Systemstudio uses the createrepo version to determine createrepo features such as as "--update" and "--database", among other things. Previously we determined the createrepo version from the installed rpm (rather than from the program itself) to work around a specific legacy issue where createrepo reported an incorrect version for the version in which the 'update' capability was introduced. Unfortunately, the current approach (obtaining the version from the rpm) causes trouble when running systemstudio in chrooted environments (like our sstest machine environments) in which the version of rpm on the host may be incompatible with the rpmdb in the chroot.

This commit addresses both chroot and legacy issues by 1) obtaining the createrepo version from the program and by 2) providing a special case to determine the "update" capability by looking for the string '--update' in the output of 'createrepo --help'.