removing boot.cat from list of installer files to download
This file was mistakenly added to the list of installer files to download for anaconda version 13.21.20.5-1. It is not necessary as systemstudio automatically generates boot.cat files when creating iso images in the iso and bootiso events. Additionally, trying to download it creates unnecessary errors in cases when the file is not present.