removed gpgsign module
Previously, when the optional gpgsign element was included in a definition, systemstudio resigned *all* rpms with a key provided by the user. This was a time consuming operation, which provided little added security in systemstudio scenarios. Basically gpgchecking is an approach targeted to alert users on client machines when rpms are provided by new sources, and ask them to verify trust. For systemstudio scenarios, installation and updates are generally performed in an unattended fashion; thus there is no "user" to verify the rpm source.

For this reason, the perferred methods for ensuring valid content is provided to client machines are 1) to limit write access to the download server 'local-dir' and 2) to configure and require SSL on the download server 'remote-url'

Thus, given the performance cost, the limited benefit, and the ongoing costs of testing and maintaining systemstudio modules, this commit removes the gpgsign module and related documentation.  It also provides documentation on the methods listed above for ensuring valid content distribution.