consistent naming for kickstart files ('ks.cfg')
This commit standardizes naming for distribution kickstart files as 'ks.cfg'. Distribution builders can still provide a kickstart file with any name, but when the file is included in the distribution systemstudio rename it. This simplifies deployment by allowing separation between the person who creates a distribution and the person who deploys it. The deploying person need only know that if there is a kickstart for the distribution, it will be located in the install tree with the name 'ks.cfg'.