no longer generating manifest file
Previously systemstudio generated a proprietary 'manifest' file, stored in the os folder, which contained md5 information for each file in the distribution. Since then, anaconda introduced a 'treeinfo' file which, when combined with rpm information in the repodata folder, performs the same function in a standard way. Thus, we are removing the file.

Removing the file provides a small performance win as we no longer need to diff changes to every file in the distribution (hundreds or even thousands of files) to know when we need to recreate the file.

Related to removing the file, we made changes to the iso event, which was using it to determine when iso's need to be recreated. Now iso tracks changes to files in the SOFTWARE_STORE. This is a slower operation, but overall a reasonable performance tradeoff since iso is an extension (optional) event.