fixed createrepo when building el5 distributions on el6
EL6 versions of createrepo began using sha256 to create checksums for repodata files. This creates trouble for EL5 versions of yum which do not have support for sha256. Fortunately, newer versions of createrepo also ad support for the --checksum option which allows specifying the 'sha' (really sha1) checksum type, which el5 versions of yum *do* understand.

This commit adds support for the --checksum option (if supported by the createrepo running on the build machine, i.e. the build machine is running el6) and sets the value appropriately ('sha' when building el5 distributions, and 'sha256' when building el6 distributions).