add try/finally around depsolve incremental update
This commit adds a try/finally block around depsolve incremental update actions. The finally block calls end on the progress timer bar. This is needed to close the status bar (which runs as a daemon process) in the case of an exception. Failing to close the status bar leaves it running in the background, which can cause errors after the main program exits, since it tries calling back to objects in the main program which no longer exist.

This commit also reverts a previous attempt to fix the same symptom (progress bar errors after program exit). That approach was to call the join method on all progress bars after they were started. It had the unintended consequence, however, of causing the timers to stop working.