removing --clear-cache command line option
This option was confusing - did it clear the shared cache or solution cache, or both? If it only cleared the shared cache (as was the case), why was a definition also required on the command line? Should it only clear the cache? or clear the cache and then continue building the definition? Rather than addressing these usage issues at this time we are removing the option. The manual alternative is "rm -rf /var/cache/systemstudio/.cache"