don't remove md5file on exit of config post scripts
Previously config rpm post scripts had been removing the md5file on script
completion. This was done as a script debugging convenience. The effect was to
cause files to show up as "changed" the next time the rpm was installed. In
theory the md5sums file should only have been deleted if the post script
failed, but the way it was implemented, the file was deleted when the post
script exited, whether success or failure.

The trouble with deleting the md5file is that the postun script relies on it to
determine files for removal. If a file does exist in the md5file of an updated
rpm, uninstall of the prior rpm removes it. Thus, the prior behavior was
causing erroneous file removals.

This commit removes the prior code. Users will need to manually add files to
the changed variable when debugging scripts that rely on it. Fortunately, now
that Deploy supports multiple config rpms, use of the changed variable is less
common than in times past.