fixed a bug that was causing updated deploy scripts to be ignored
Previously fileio was tracking differences in text-content files using the
config handler. There was a bug, however, in that it was storing a relative
xpath to the content, but looking for changes using an absolute xpath. As a
result differences were never found. This bug was evidenced in deploy modules,
but not in config-rpm modules, as the latter regenerate ALL files and scripts
any time ANY change is made to config data.

Tracking file content using the config handler had another issue, which is that
the xpath may contain positional identifies (e.g. file[1], file[2], etc.). As a
result fileio would unnecessarily regenerate a file if its position changed
within the definition, but the content remained the same.

This fix handles both issues. Fileio now tracks differences using a variable
keyed by the destination file path. And it uses the same path for both
storing data and identifying differences.