repo/download now defaults to 'false' for all repos
Previously we had special case logic for repos with the id 'base' and
'updates'. These repos defaulted to 'false' and all others defaulted to 'true'.

To make repository definition easier, we are going to move away from
the standard 'base' and 'updates' repoids, however. Instead we will use
ids that incorporate the os version and edition, e.g. 'rhel-server-6'.

Because the common case for package repositories is to *not* include
external packages, the cleanest way going forward is to default to false
for everything.