update ks.xml and documentation to include all packages '*'
Once upon a time, we included all packages using '*'. Then for some reason in April 2015 we changed to including just the core group. I think this is because we assumed the core group could be installed, and additional packages could be downloaded using the update script. This approach wasn't working properly on el7 systems (for some reason, I didn't investigate closely), and left the system installed with just the core group and no access to the system repo where other packages could be downloaded. In a pragmatic step to work around this, at least for now, I changed the kickstart to include all packages again.